No longer import db_api while parsing db __init__.
The db_api field of reddwarf/db/__init__.py was grabbed by importing a
module using utils.import_object. The imported module was always the
sqlalchemy api module, which is nested under the file that imports it,
which caused import errors in multiple contexts. This commit changes
every use of db_api to call a function to load it first, which seems
to fix these problems.

Change-Id: I9515efe02831b521d7f14674e0a2913e476b207d
Fixes: bug #1080706