Better defaults that fix volume support
The default value for block_device_mapping should not be a path, rather should be the name of the actual device in /dev.  The device_path should also be defaulted to the fully qualified path to that device.  This patch fixes volume support and does not require you to set reddwarf_volume_support=False to get a guest to boot sucessfully.

fixes bug: 1084802

Change-Id: I8aa8582da7b400c91f761a2974f859df8cd45003