Adds optional hostname to calls involving users
All calls that involve calling a user by name now
also allow for the host to be specified optionally.

Similarly, all calls that respond with a user also include
the host, defaulting to '%', MySQL's shorthand for 'anywhere'.

There are also a few fixes to get all real, fake, and tox tests
to work. Those will probably be rebased away if other reviews are
merged first.

Change-Id: I6df821ef2ad7d0602a060c853f0ef9bdc5c11c26