Adding checks when updating a quota
The rd-client is changing to no longer accept all resources when updating a quota.
The client will then send 'None' values to resources that are not to be updated.
This fix ensures that the changes in the client are handled.

This effort was initiated by the implementation of quota for backups

Change-Id: I24e02a2736750a2606d62dfd06f0130bef503b10