Fixes test in response to optional hostname fix
Several tests have had their behavior changed
with the client appending the default host.
With that new behavior reversed in the client,
some tests use a username that contains an @,
causing an ambiguous case for the API to parse.
Deliberately including the host in the tests
fixes this.

Fixes Bug #1171559

Change-Id: I967eb4b21d8ca3e2c2721db6f362b8dac2551346