Don't require an admin user if there isn't a need.
This change makes it possible to skip specifying if a user has to be
an admin or not as in some contexts it may not matter. Found while
trying to run the tests against an environment where my Nova user
worked only for the Nova MGMT api.

implements blueprint tests-should-not-require-admin-prop
Change-Id: Ief62d47216b6366e36fd90e86baf81172e5281ba