Fix faulty 404 errors when requesting bad versions
The problem came down to the `_dispatch` method of the openstack
common wsgi router object. When a match was not found it returned
a webob 404 exception directly rather than allowing it to be
wrapped in a serializer.

The fix involved extending the `Router` object and reimplementing
the `_dispatch()` method to use the `Fault` object that serializes
exceptions.

Change-Id: I24a590f65ff655b25cfd7d84786df3055af701f1
Fixes: bug #1174960