Support Security Group Name Prefix Customization
As of now, if CONF.trove_security_groups_support == True, then a
Security Group is created for every new Instance, following the
naming scheme of: "SecGroup_<Instance's UUID>".

Nicira NVP for Quantum enforces a Security Group naming limitation of
40 chars maximum. Given the UUID's length, this only leaves 2
characters for the prefix, e.g. 'SG' vs. 'SecGroup'.

Even in lieu of restrictions, it's not inconceivable that a
vendor/user might want to customize the naming scheme of their
Security Groups to align with other cloud providers, etc.

Change-Id: I1043e15c71607cabe2fd6f72f64705e80cd2cde1
Closes-Bug: #1218589