Fix checksum verification exception.
Previously if an invalid checksum was found, the exception would not be
raised. This commit also makes the tests for backup run in fake mode
run again (turns out there was no reason to disable them) since they
helped find this, and cleans up an issue with the restore backup test
(it used the current size of an instance created elsewhere for no
reason).

Fixes bug 1246368

Change-Id: I96a570a93695b017e5c6966fa1482d6ec5aef426