Extract suffix from req URL to avoid escaping dots
Usernames and their hosts can contain dots. As such,
user CRUD requests may end in ids that contain one
or more fullstops. Our request router can interpret
these as file suffixes and our methods and as a
result, truncates these ids improperly.

This change inspects the request object in these
cases, looks for a possibly misinterpreted suffix,
and reattaches the suffix to the id. This means
that escaping . in requests shouldn't be necessary.

Closes-Bug: #1252327

Change-Id: I74c4d6694baba44ebe0018eedac2ae5c4773b370