use_stderr should be False for tox tests
Reverting the change to the trove.test.conf so that the tox tests are
not so verbose with all the logging when running through the tests. It
makes it very difficult to find out what failed and where without much
groking of the output.

This was a change i made so that when you do a './redstack start-fake'
it would show the logs in the screen session. But this turned out to be
a bad idea when running in tox.

Change-Id: I236b2de2e89e49fd937d917d2955fddb5fb35d6f