Fixes resizes for volumes attached to active Nova servers
Currently, Trove fails to resize volumes attached to active Nova
servers. It also skips crucial steps such as unmounting and
mounting the volume and resizing the filesystem. When extending a
volume attached to an active Nova server, the guest must be
stopped, the volume unmounted, detached, and extended. Once the
underlying volume has been extended the volume should be reattached,
have the file system resized, mounted, and then the guest should be
restarted. If all operations complete successfully the task status
should be set to NONE and a usage event sent, reflecting the size of
the volume as reported by Cinder.

This fixes extending volumes attached to active Nova servers to ensure
a proper resize is performed and the task status is set to NONE and a
usage event is sent after all resize operations complete successfully.
Unit tests are also added.

Change-Id: I9a5785f539a2ee138ce1ded3cbd795ef91d66400
Closes-Bug: #1259976