Instance details view shows hostname (if it has it) or IP
Currently the hostname and IP is controlled by two config values:
trove_dns_support and add_addresses. I dont see a point of having
the hostname and IP indenpendent. It causes issues for old instances
when DNS is enabled.

I propose to simplify the view by showing hostname if it is set or
show the IP if it is not set.

Closes-bug: #1260567

Change-Id: Ia1fd47cbee89b10d53b0821e104a34eb31543846