Add Neutron support
Currently create instance doesnt work in OS installation with Neutron.
To get it work, additional parameter 'nics' should be specified
in Nova 'create' call.
This change allows user to pass 'nics' parameter when create instance.
Also possible to specify list of network IDs which should be attached
to instance if networks are not specified by user.

Closes-Bug: #1257838

Change-Id: Ifd98a880dfa22e9677e802bcd779acd9146ec671