Add Volume Resize filesystem size assertion
Verify that the filesystem size on the volume increases and is about
the size of the Cinder volume.  This checks for the case where the
Cinder volume device may have increased but the filesystem may not
have been resized thus not representing the actual available volume
size.

Change-Id: I06059cd18ec78085187c4e86ea9288d35ee4a5f6
Implements: blueprint add-volume-total-attribute