Fixes a race condition in resize flavor
When resizing an instance's flavor, Trove asks the guest to start the
datastore via an RPC call and then, to be safe, would check the
service_status table in the database to make sure it was updated to
RUNNING. The thing is, that second check was not only superfluous, it
was causing resizes to fail in the post-conductor world, because while
the guest would have sent a message to conductor describing the
datastore's current state before it informed taskmanager that it had
finished the call, conductor itself might not have updated the
database. This commit changes things so taskmanager polls until the
service_status is RUNNING after calling the guest. If there is an error
the call to the guest will fail anyway.

Change-Id: I4e1ca75a150ed58233c21372d21c6337596e43d0
Closes-Bug: 1278282