Fixes a race condition in resize flavor for Percona
When a resize happens immediately after an instances gets created,
the mysql process doesnt have enough time to finished up
provisioning before the resize-action invokes a stop and start
resizing. Although the trove tries to stop mysql, the stop command
doesnt ensure that mysql was stopped successfully. This causes
mysql to be corrupted when migrated to a different VM. Normally
this is not an issue since mysql provisioning has plenty of time
to finish. This becomes an issue when doing an resize when mysql
has just finished provisioned. Only happens in automated tests.
This fix will ensure that the mysql server is stopped before a
resize is performed.

Change-Id: I0a3c351b61b82b88dfe05325a38ce93d6973a9f6
Closes-Bug: 1284989