Add support for 'trove root-enable' in Couchbase
Currently, trove api call 'root-enable' is not implemented by Couchbase
guest agent. This fix adds the required support for the api call.

Root-enable api call is being handled by the root controller in mysql
extensions module. The controller makes a sync call to the guestagent to
enable the root. The guest agent handles the actual implementation. This
being true, the api call should be generic for datastores and not a
part of mysql extension module. If a particular datastore does not
support root-enable, the guest agent returns an appropriate message. The
abstractions in the extension module are not clear and blueprint should
be filed to generalize the extension module.

Closes-Bug: #1294821

Change-Id: I3f098032333cda70f4181e5688514cbda6730a45