Fixed unit test to not actually run backup command
One unit tests is still trying to execute the actual backup command
instead of the mocked backup command. This is causing an issue on
the Jenkins slave where it is trying to run as sudo (and failing).
Fixed the test to use a mock runner, and not try to run innobackupex.

This had been previously fixed by commit 19d940516e, but that fix was
clobbered by the move from mockito to mock.

Change-Id: Ib63187d4414a3ef7b7f664e076bb4d92a40c2214
Closes-bug: 1291516