Move sql_xx params to [database] conf section
Update trove configuration files to use the current convention for db
options:
- deprecate the used of [DEFAULT] sql_connection, sql_idle_timeout,
  sql_query_log
- add support for [database] connection, idle_timeout, query_log
- update sample/test conf files to use the new convention

DocImpact: Change to conf options in Install Guide - Install the
  Database service.

This has become more important, as the recent changes to devstack
now use the new [database] section, and the gate is failing due to
that.

Note: On my quick look at the code, there's more work needed to use
oslo_db properly, and it might be better to do that out-of-band
of this fix.

Authored-By: Greg Lucas <glucas@tesora.com>
Co-Authored-By: Peter Stachowski <peter@tesora.com>
Closes-Bug: #1308411
Closes-Bug: #1434856
Change-Id: Ic9e5b20e21d09c4b1d49df660b89decf094440bb