Datastore_versions.name shouldn't be unique
The name column in the datastore_versions table is not supposed
to be defined with a unique index. The migrate version 18 was
intended to fix this but the method used to correct the column
definition doesn't work. Added a migrate version 26 to use
UniqueConstraint.drop() instead.

After some discussion, it was decided that downgrade would be
a no-op because it is difficult to reliably create a unique
index on a column with data in it, and further, the column wasn't
supposed to be unique in the first place.

Change-Id: Ida552982de6702db6ed4d40c0167e7d542c0ded4
Closes-Bug: #1310834