Remove admin_token from configs
A previous review removed the admin_token from the sample config file.
It appears that devstack doesn't use that file but instead uses one of
the .test files.

Admin tokens are actively discouraged as a form of authenticating in
auth_token middleware and the behaviour when using admin_token in
combination with a username and password (as this default produces in
devstack) can lead to problems in authentication.

Closes-Bug: #1299332
Change-Id: I78d2c184f7937006b808d580e3663ab055f7c5fe