Fix enable on boot when working with systemd and symlinked units
mysql on Fedora20 is actually mariadb and /usr/lib/systemd/system/mysqld.service
is a symbolic link to /usr/lib/systemd/system/mariadb.service. systemctl "enable"
doesn't work on symlinked inits like described in:

https://bugzilla.redhat.com/1014311

So in order to enable mysqld service properly in this case we need to
replace a service name with the real path to service file.

Closes-Bug: #1330937
Change-Id: Id823b3e377c4d0d5f1e5452fad9826232bc1f904