Use (# of CPUs) api/conductor workers by default
This changes the default number of trove API and conductor workers to be
equal to the number of CPUs available on the host, rather than
defaulting to 1 as it did before.

Commit 75c96a48fc7e5dfb59d8258142b01422f81b0253 did the same thing in
Nova in Icehouse. Similar changes are being made to Glance and Cinder
as well.

DocImpact: trove_api_workeres and trove_conductor_workers will now be
           equal to the number of CPUs available by default if not
           explicitly specified in the trove configuration files.

UpgradeImpact: Anyone upgrading to this change that does not have
           trove_api_workers or trove_conductor_workers specified in
           the trove configuration files will now be running multiple
           API and conductor workers by default when they restart the
           respective trove services.

Closes-Bug: #1335284

Change-Id: Id300bbe991436a0f826ea715630669ab5922a6a4