Stop cassandra during configuration and volume migration
The cassandra service is left running after installation.
The next operation(s) are to rewrite the configuration, and
migrate the data onto a volume (if provided). It is not a
good thing for the database to be running during the latter,
and a restart is required after the former.

Failing to stop the database during the data migration can
result in an rsync error 24 because files that were present
when rsync made its list may be missing by the time rsync
gets around to trying to sync that file. This can lead rsync
to not make a 'true' copy of the source on the destination.

The only way around this is to stop the database, something
that is already done by both MongoDB and MySQL in their
guest agents.

Change-Id: I74ecc5c12d43eccc6295488e344bc0e75fe8116d
Closes-bug: #1338660