Logging audit for trove/mysql module
Adjust logging to conform to logging standards.
Cleaned up a few messages that are unclear and
changed some LOG.info messages to LOG.debug.

Also fixed a pattern of having an exception with
the sole purpose of logging an error with the
value thrown and replaced with LOG.exception
instead.

Added LOG.debug messages to the mount, unmount
and resize methods in service.py; this change
was made to all datastores.

Change-Id: If2a5c95b806e555d077ce494f474d3463fda6794
Partial-Bug: #1324206