Handle exception from pexpect child.close
When pexpect child.close(force=True) is called to close a sudo process, an
exception may be raised on certain operating systems. Terminating a sudo
process as a non-root user succeeds on Ubuntu 12.04 but fails on 14.04.

Handle this exception and try to kill the child process using 'sudo kill'.

See related change https://review.openstack.org/#/c/100353 that addresses
another instance of the same issue.

Change-Id: I4bd9e4b98479e33604545713337440ce81ca64ee
Partial-Bug: 1343344