Document Trove configuration options
Many of the Trove configuration options do not have adequate help text
to describe what the configuration option can be used for, and/or what
valid values for the option are.

Since the auto-generated docs are based on this help text, they
currently look very empty and are not very helpful.

example: See "Configure the database" in chapter 4. Database Service
         under http://docs.openstack.org/trunk/config-reference/content

All options now have help strings to clarify how/when they can be used.

Also removed periodic_interval option (not referenced in the code).

DocImpact: A separate patch set will be created for the actual doc
changes.

Closes-Bug: #1334465
Change-Id: I3ea1bd255514438c3d356d23d42f3fb3ddd45274