Set the python hash seed that tox uses to 0
The current trove unit tests don't work correctly with the random
hash seed that the new tox sets by default.

This change forces the new tox to use 0 as the hash seed which
should unblock the gate until we can fix the trove unit tests
to run correctly with a random hash seed.

This change will probably need to be backported to the stable
branches to ensure that the tox tests keep working there as well.

Change-Id: Ib5c6d4d166c847ce94ec70740c5753d338687865