Complete mocking for test_extensions
The current implementation of trove.tests.unittests.api.
common.test_extensions does not fully mock the
test_default_extensions method. As a result trove UTs run without
trove being installed fail in this method as it assumes the
entry points have been installed into the egg for trove.

This patch completes mocking of that method so that the
test_extensions class can be run in a UT environment in
which trove is not installed and thus may not have the entry
points installed to the egg.

Change-Id: I2543f17fe087cf98d1a11ef47a07b4addfd32cd4
Closes-Bug: 1372950