Increase test rate limit to avoid rate limit error
The current rate limit is 200 requests per minute. I'm getting
consistent tox failures as the machine is exceeding this rate. The fix
impacts only test code and adjusts the limit and a test that has a
hard-coded reference to the old limit. Why 500 you may ask? Because
600 worked and 450 failed consistently with the rate limit error.

In addition, the change addresses the fact that some test
configuration values are duplicated in the test; the change makes the
test reference the configuration value.

Change-Id: I4bb290d8de6253d65b7877c743bb288ee2bce536
Closes-Bug: #1376689
Closes-Bug: #1378932
(cherry picked from commit a67c0ca7030362c35ea8dd183229bc73173cf60b)