Inject guest conf files to configurable location
Introduce a configuration option for a directory on the guest where the
taskmanager should inject configuration files.

During instance creation inject the guest_info and trove-guestagent
conf files to the 'injected_config_location'. The default location is
/etc/trove/conf.d.

Also:
 - Change the default value for 'guest_config' to remove the risk of
   overwriting a guest image's conf file with the sample.
 - Add trove-guestagent.conf injection to heat template.
 - Add new 'guest_info' option that defaults to "guest_info.conf".

Depends-On: I1dffd373da722af55bdea41fead8456bb60c82b2

Co-Authored-By: Denis Makogon <dmakogon@mirantis.com>
Co-Authored-By: Duk Loi <duk@tesora.com>

DocImpact: This change introduces a new option with a default that
affects existing guest images. The guestagent init script for any
existing guest image will need to be modified to read the conf
files from /etc/trove/conf.d.  For backwards compatibility set the
injected_config_location to /etc/trove and guest_info to
/etc/guest_info.

Closes-Bug: 1309030
Change-Id: I1057f4a1b6232332ed4b520dbc7b3bbb04145f73