Obsolete oslo-incubator modules - unused modules
This change is part of a multi-part change set to handle obsolete and
graduated oslo modules.

This commit handles obsolete oslo-incubator modules that are not used
in Trove but are referenced in openstack-common.conf

The change here is to remove them from openstack-common.conf and
delete them from trove/openstack/common.

While excutils, network_utils and strutils are deleted from
openstack-common.conf the files are retained because of dependencies
within oslo-incubator.

Change-Id: I77535f6798f202f51f04b94ee9271b139b5c6b78
Blueprint: retire-unused-oslo-incubator-modules
Partial-Bug: #1380789