Fix timeout in test_slave_user_removed int-test
The test_slave_user_removed int-test polls to check that the
slave user account is removed, but doesn't currently have a time
out set. This means that a failure in this test causes the entire
suite to spin, and ultimately time out. Fix this by adding a 3 min
timeout to this test.

Change-Id: I5f869c7d3cb0587409b5a12a89ed673e57552b82
Closes-bug: 1399583