Obsolete oslo-incubator modules - timeutils
This change is part of a multi-part change set to handle obsolete and
graduated oslo modules.

This commit handles the timeutils module.

The change here is to remove it from openstack-common.conf and point
all modules who used to use that at oslo.utils. This also involves a
change to requirements.txt.

I have rebased this change and made it dependent on the change in
https://review.openstack.org/#/c/129292/

Change-Id: I31176c0a78e734bfad3b40b2ec634ea8f362d81b
Partial-Bug: #1380789
blueprint: retire-unused-oslo-incubator-modules