Obsolete oslo-incubator modules - jsonutils (now oslo.serialization)
This change is part of a multi-part change set to handle obsolete and
graduated oslo modules. This commit handles the jsonutils module which
is now oslo.serialization.jsonutils. The change here is to remove it
from openstack-common.conf and point all people who used to use that
at oslo.serialization. requirements.txt has been updated to point to
this dependency.

The file jsonutils.py is not being deleted because of dependencies in
oslo-incubator.

Change has been rebased on https://review.openstack.org/#/c/129294/

Change-Id: I2a7d4c41ce77d6375fa08033514baa815a087016
Partial-Bug: #1380789
blueprint: retire-unused-oslo-incubator-modules