Obsolete oslo-incubator modules - processutils
This change is part of a multi-part change set to handle obsolete and
graduated oslo modules. This commit handles the processutils module
which is now oslo.concurrency (oslo_concurrency). The change here is
to remove it from openstack-common.conf and add oslo.concurrency to
requirements.txt. Since nothing in oslo-incubator depends on
processutils, the module is also deleted from the oslo-incubator
imports.

This change is based on top of https://review.openstack.org/#/c/133068

Change-Id: I5cba2f487b51c877d036acf552df3b64f7c85764
Partial-Bug: #1380789