Correct a test and order of parameters to assertEqual
In debugging a failing test (test_update_heartbeat) I observed that
the test was asserting incorrect values and also that the order of
parameters to assertEqual were inverted. It appears that the order of
parameters to assertEqual is a more widespread problem.

The correct order makes the assertion message more useful.

Change-Id: Ide9f60f8338398ad8e61ad81f4a30058e415f8e9
Closes-Bug: #1424741
Partial-Bug: #1277104