Fix replica source state validation
Creating a replica did not check that the specified source was available.
It is necessary that the source's state is active and that it is not busy.
This is done by verifying the state and status of the master.

The new checks will throw a HTTP 422 UnprocessableEntity in cases when the
status is not ACTIVE or source is busy performing a task

Changes:
 - adding validation step that checks if master exists and in proper state.
 - added appropriate API and unit tests

Co-authored by: Denis Makogon <dmakogon@mirantis.com>

DocImpact

Change-Id: Ib4e37ff0036998bc53058e400054ed93fcc2e144
Closes-Bug: #1357704