Add support for DB2 datastore in Trove
This feature will provide support for a new SQL datastore in
Trove. Users will be able to provision a DB2 server instance
and perform various functions as listed below. The risk of
regression is very minimal since there is no code changes to
the base Trove code. The guest agent for DB2 uses the existing
API's supported by Trove to provide the functionalities
mentioned below.

The following features have been implemented in this patchset:
- Launch
- Reboot
- Terminate
- Resize Volume/Instance
- Create/Delete/List database(s)
- Create/Delete/List user(s)
- List Access
- Included unit tests for the different methods

DB2 Express-C can be installed on a server that satisfy the
memory requirements of 512MB to 4GB of RAM.

Another point to note is how DB2 creates and deletes users.
There isn't a notion of a DB2 user in DB2 unlike in the case of
MySQL. When we create a user for DB2, the process involves first
creating an OS user and then granting privileges to this user for
the given database(s).

Currently all the DB2 queries are run from command line. One of
the enhancements that can be made is to use the DB2 SQLAlchemy
driver to run these queries. This requires the guest agent to know
about the username and password to connect to. We can use the
'db2inst1' user (which is the instance user for the DB2 server
instance created) but since DB2 does not track the passwords, there
isn't an easy way to get the password unless we stored it in a
file or use a similar mechanism. Since I am not sure how safe that
is, I decided to go with running the commands from command line.

Change-Id: If95686dd37d2730a4d00f4010fa8515f33a8e337
Implements: blueprint db2-plugin-for-trove