Fixes config templates for mysql & percona
Calculations for certain config parameters was resulting in zero
if a value < 512 was used as ram for flavors to boot guest instance.

Replaced '//' by '/' and the result is again casted to integer value
as needed by mysql configurations.

// implies divide two numbers and return
the truncated integer result.

whereas

/ implies divide two numbers and  return value
will be a floating point number.

Updated default instance configuration response json.

Change-Id: Icfc84a4cb380264670aa4e31f372eb39a6e87c7c
Closes-Bug: #1440358