Remove [Service] section from mongo config tmpl
The mongodb guest wasn't starting up on Ubuntu or Fedora
due to the presense of a [Service] section in /etc/mongodb.conf.
I believe this section was not intended for config.template but
should be in the systemd script for mongo. For now, we will
revert this change and consider a better way to fix it for
Liberty.

The change that introduced this is:
https://review.openstack.org/#/c/150131

Change-Id: I090c95185f732485a21f6d38df1655e205a07224
Closes-bug: 1446735