Fixes the unsafe mocking in test_dbaas
test_dbaas.MySqlAppMockTest.test_secure_with_mycnf_error
contains a mocked entry for operating_system.service_discovery.
This entry is unsafe and it has not been unmocked, also

It looks to me the primary reason for this bug: 1450789.

Made the mocking safe by the usage of "with patch.object(..." usage

Change-Id: Ib0f18b65be04490f8f60ff878a7af512a5e9bed5
Closes-Bug: #1450789