Move guestagent settings to default section
The guestagent related settings are in the database section of the
taskmanager.conf.sample file.  If a user were to use this sample then
the guestagent related settings will not be read by the taskmanager
service because it expects these guestagent related settings to be
in the default section.  This fix moves the guestagent related
settings to the default section.

Change-Id: I8dbdf7abcfd5f2341831cc6daaf5cb14d8d42d61
Closes-Bug: #1459415
(cherry picked from commit 9b47609941d338da11e29d79f4a74e6b3bdd6e3b)