Provide option to read SSH credentials from test env
Currently the SSH user and private key is assumed to be taken
from the local account (e.g., /home/ubuntu/.ssh). This change
proposes to add 2 environment variables so that a tox test of
Trove can use an arbitrary SSH user and key file for testing.
Obviously, this assumes that the test setup has already
installed those credentials in the image under test.

Change-Id: Ibc83cb7fcd291a13706550965fe2dd1d5ca3504e
Closes-Bug: 1469775