Added replica_of attribute to test_index_list
The test_index_list made an assumption that you are running
trove on a clean environment without any previous instances
created. If you happen to have created a replica of an instance
the test would now fail. Adding 'replica_of' to the allowed
attribute will allow this test work with replica instaces.

Change-Id: I80ec0e5f26ea52339082d540f2747e54f08e279b