Percona Xtradb Cluster implementation
Created the clustering implementation for Percona XtraDB Cluster
Server. Used the existing framework for adding a new cluster strategy
for percona. Created guestagent, api, and taskmanager calls to setup
the cluster.

Ran into a weird issue where creating the admin user was not being
saved even though there was no error logged. I believe this was
happening because mysql was technically up from the mysqladmin ping
cmd but was not processing the grant call yet. I've heard that to be
sure you should run a 'select 1;' call to verify that mysql is really
ready after starting it. I added a method to verify this.

PXC needs a few more ports opened up for the group comm, state
transfer, and incremental state transfer. Added the percona
cluster.cnf template file for the setting needed in PXC. When a
joining instance first joins the cluster it syncs all the data using
xtrabackup, this will cause the joining instance to lose its admin
user os_admin. So we reset the admin to a new password across the
board.

Make Percona Xtradb Cluster a different datastore because there are
muliple version of PXC. This has been tested with pxc 5.5 from the
ubuntu repository.

Adding coverage xml output in tox.ini

DocImpact
Adding a new datastore percona-xtradb-cluster(pxc) that should not
have any special cases for create and delete cluster.

Implements: blueprint support-pxc-56
Change-Id: I239433f0491cea15b214c41ceecc3a5e72b5bbeb