Add Redis backup/restore functionality
Performs backup using the Redis client to persist data to the file
system, then streams the result to swift.

Performs restore by replacing the data file with the Swift backup
and starting the server again in the correct manner.

Note: Running the int-tests require that volume_support is set
to false in the test.conf file.

To run:
./redstack install
./redstack kick-start redis
(vi /etc/trove/test.conf and change volume_support to false)
./redstack int-tests --group=backup (or --group=redis_supported)

Co-Authored-by: hardy.jung <hardy.jung@daumkakao.com>
Co-Authored-by: Peter Stachowski <peter@tesora.com>

Depends-On: I633273d438c22f98bef2fd1535730bcdb5e5cff0
Implements: blueprint redis-backup-restore
Change-Id: I1bd391f8e3f7de12396fb41000e3c55be23c04ee