Add support for MariaDB datastore in Trove
A specification for this change was submitted and reviewed in
https://review.openstack.org/#/c/211330/

This patchset adds the hooks and configurations options required.
No operation has been reimplemented, MariaDB 5.5 behaves exactly
as MySQL 5.5. Because of this, this patchset depends on the MySQL
Manager Refactor, which makes it possible to the MariaDB driver to
inherit the implementation of the MySQL driver.

Implements blueprint mariadb-datastore-support

Change-Id: Ia1d77fee82b08b24f5f379ee1534f335db97fd57
Depends-On: I633273d438c22f98bef2fd1535730bcdb5e5cff0