Fix Postgres services management
Use the existing shared operating system interfaces to manage the
Postgres service.

Use the 'state_change_wait_time' configuration value
instead of the currently hardcoded '30s' as start/stop/restart timeouts.

The existing postgres implementation does not wait for the
database to change its status after a service call.
This behavior can potentially lead to severe transient issues,
such as FS corruption (unmounting the volume when the database is
still writing to it) etc.
This patch fixes the above issue by making sure to always wait for the
database to actually start running or shutdown before proceeding.

Abstract the start/stop/restart database calls to tested reusable
interfaces of the Status class, so that all datastores
can make use of them rather than having to re-implement and test this
common functionality from scratch.
It should hopefully help to cleanup guest-agent implementations and
prevent the above issue in the future.
NOTE: The added calls would be good candidates for base-class
methods once we have a base service class for the guest-agent.

Closes-Bug: 1492526
Change-Id: I0ab198be4f96e3ba3c2e8a158d11de3a44046ab3