Properly patch _init_overrides_dir in MongoDB
Since most of us don't have MongoDB installed on our
desktop/development environment, tox tests will fail because some
config paths for Mongo aren't available.

The change proposed here is merely to mock the _init_overrides_dir
properly in changes that would otherwise stub their toes on this.

It is my understanding that this issue was introduced in Change-Id:
Iaca9cc9e3ce1e6521839062fc9458c1b987f9464 which sought to address bug
1485158.

The reason this problem occurs on my machine (and likely on other
developers workstation) is that I don't have MongoDB installed. The
reason this doesn't impact the gate is that those machines have
MongoDB installed (for ceiliometer).

Change-Id: I0f3d24b95a7d2ef1e23a0f1fd8f60bc496261f70
Closes-Bug: #1495573