Fix promote for Redis datastore
The 'set' command in the Python client interpretes None as the
string 'None.'  If no password is set on the master, a detach/attach
(such as is done with promote) will have the slaves try and use 'none'
as a password.  This will cause them to not be able to connect to the
master and leave them in a detached state.

The 'masterpass' is now set to an empty string if no password is set.
The restart of Redis was also moved before attaching the slave,
so that the mapped CONFIG command is recognized properly.  This is
required so that setting the master password completes successfully.

(cherry picked from commit 6c4db5d327df7e10f7a0bf38deaee0db4fe2ee5a)
Closes-Bug: #1498208
Change-Id: Ibc842b95fd191908d4fac420030f84393c696013
Depends-On: Ida0aa155b1e1ec7b8e9fa73006d3db9a88ffea02