Refactor the datastore manager classes (more)
This moves most of the rest of the boiler-plate code
into the base class.

This includes more filesystem methods, the 'unsupported'
methods and a configuration method.
It also consolidates the replication strategy.

The prepare_completed flag was also prefixed with a
double underscore to make it difficult for descendant
classes to change it by accident.

Log messages were also enhanced.

Partially Implements: blueprint datastore-manager-refactor
Change-Id: I2010ffd8e8f480b12b0cfd30ffc948a7c8963bd8