Cleanup trove debug and info logging messages
Initially this was related to the case of the conductor where
consecutive lines had debug messages that I was interested in, in
debugging a problem. When there are more than one guest instance
running, multiple threads in the conductor causes these messages to
get intertwined. Which is a "pain".

In fixing this, I found some other problems, such as that
QuotaController had logging using the _ i18n function but didn't
actually import i18n (a problem waiting to happen) and that some were
missing useful information. These have been fixed as well.

I've addressed those as well. For the most part, this change replaces
multiple consecutive lines of logging at the same severity level with
a single line that creates a dictionary with all the information for
the message.

I did a general sweep of the code to look for cases where we used
LOG.debug with a message that was i18n'ed and for cases where we used
one of the other LOG. functions without an i18n'ed message and fixed
those as well ...

Change-Id: Ia6d6b79d431291ca667d1502b6b073ec91ac5f7c
Closes-Bug: #1507290