Correct the computation of elapsed time while waiting for state change
The computation of elapsed time and present time during waiting for
state change is incorrect. This change simplifies the code and
corrects that error.

There is a subtle change in behavior also introduced as part of this
change. Earlier code would unconditionally wait for WAIT_TIME before
checking status. The proposed code will immediately poll status.

A configuration parameter has been added for the WAIT_TIME with the
former default value of 3.

The change(s) to test_dbaas (the tests) are because of an unfortunate
side-effect of the old code. The tests would just mock sleep() but
the old code would advance the amount of time waited by whatever time
they thought they'd have waited for. This meant that the
wait_for_real_status_to_change_to method would take approximately no
time in tests but would appear to do all the right things. However,
since the new code is using time.time() to deterime how long it's been
around has caused the mock'ing of sleep() to have an unexpected
result; the wait_for_real_status_to_change_to method becomes a CPU
hog! The hack I've used is to fake time with a side_effect that
advances at 1 'second' per call.

Closes-Bug: #1286379
Closes-Bug: #1514991
Change-Id: Ibdeb94f30fee9d4337ec5095794466a61caf09ad
Co-Authored-By: Amrith Kumar <amrith@tesora.com>
Author:    Petra Sargent <petra_sargent@yahoo.com>