Correct some issues with devstack plugin for trove
The redstack script used to properly configure validation-rules if
they were available but devstack never did that. When the plugin was
created, it didn't do it either.

In testing the fix for this, it became apparent that the code for
configuring guest images was never workable; the TOKEN variable was
never really initialized.

Both issues have been addressed.

Change-Id: I6665720831337fba644a30f3ee0d5ccb77ef1e4a
Closes-Bug: #1448130
Closes-Bug: #1514951