Stop using trunk python-troveclient in test runs
Because it has backward incompatible changes like:

I5ab58fd64f4745a2e558392bff649b05401ed57a

So the change here is to use the version of troveclient
from stable/kilo global-requirements.

We also sync up with global-requirements since (1) that's a thing
that needs to happen anyway and (2) we need to cap mock to get
the unit tests passing.

Also fixes the remaining unit tests that are failing.

Co-Authored-By: Craig Vyvial <cp16net@gmail.com>

Change-Id: I465c787a65ec8cf9914c7133bd52ea8cf4741d61