Add better input checking for MongoDB
A user-create on the Trove admin user "admin.os_admin" can change the
password of that user, rendering the guest agent useless (the password
it tries to use no longer being valid).

A user-create can also be used to change the password of existing users.

The guest does no validation of inputs for various user-* commands.

None of the above should be allowed.

The fix is to skip any names on the ignored user's list and add more
checking for existing users in codepaths.

Tests are also updated as needed.

Change-Id: Icae98f57f7beda23e52b7e59a1a386c03ff20cf8
Closes-bug: 1517622