Use a specific network for the test
If integration tests are executed in an environment with multiple nics,
nova complains and the instance does not start. Given that the
shared_network parameter is already available with that specific
meaning, use it on instance creation.
The default value (None) should ensure that the default behavior
does not change.

Change-Id: I8554ebf111cad829deb34d3d23d18f72d7f13554