Move 'enable root on prepare' to the base
When given password on prepare, the 'root' should be enabled in the
base class, similarly to how the configuration overrides get applied.

It is achieved by calling already existing 'enable_root_with_password'
manager method. The call is encapsulated so that it can be easily
overriden if necessary.
The action is performed in the block where other users/databases
get created.
Note I placed user/database creation and root action into two
separate try-blocks. This is to ensure that failure in one does
not stop the other. It also allows for more accurate error message.
I made the above actions log, but not re-raise exceptions so that
the 'post_prepare' can still run in case of an error.

Datastores that do not implement root actions will have an exception
recorded in the log and the root password will be ignored
(i.e. the instance will come up without 'root' user).
This corresponds to the current behavior
(except the error gets logged rather than ignored).

Also extended the base manager test accordingly.

Related-Bug: 1532583
Change-Id: I982d94d20c9ef317d83827f898b236d0e44e95a5